[schema] clean up the implementations of schema bc check
*Motivation*

There are a lot of duplications in schema bc check implementations.

*Modifications*

Removed duplicated code.
